Updating renderer and decoupled engine and renderer
Removed the engine's methods that were only used to call the renderer's
Made the renderer be accessible from the engine without calling a method /!\ This technically allows users to modify the engine's renderer at runtime. While unaccounted for, this shouldn't break the rendering pipeline if done by capable hands.
Added possibility for render targets to be paused (should be mostly useful for resizing without recreating the framebuffers every frame